pyproject.toml
[project]
name = "langgraph-poc"
version = "0.1.0"
description = "LangGraph multi-agent POC (ADF / ADLS / ServiceNow) with LangSmith tracing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "langchain==1.2.15",
    "langchain-core==1.2.28",
    "langgraph==1.1.6",
    "langsmith==0.7.30",
    "mcp>=1.27.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.44.0",
]

[project.optional-dependencies]
adf = [
    "azure-identity>=1.17.0",
    "azure-mgmt-datafactory>=9.0.0",
]
agent = [
    "langchain-openai>=0.3.0",
]
